# My Project.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "My Project",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register_User",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "name",
                  "password"
                ]
              },
              "example": {
                "email": "a@gmail.com",
                "name": "alice",
                "password": "secret123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "email",
                    "id",
                    "name"
                  ]
                },
                "example": {
                  "email": "a@gmail.com",
                  "id": "d839872f-e8d6-480f-a76a-8f5ad5ecc82c",
                  "name": "alice"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "summary": "login",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "example": {
                "email": "a@gmail.com",
                "password": "secret123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "access_token"
                  ]
                },
                "example": {
                  "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJkODM5ODcyZi1lOGQ2LTQ4MGYtYTc2YS04ZjVhZDVlY2M4MmMiLCJyb2xlIjoiVXNlciIsImlhdCI6MTc1NTE0ODUzOSwiZXhwIjoxNzU1MTQ5NDM5LCJqdGkiOiJsQXRmLUpwOENOYjh1clFWZno5X0p3In0.ieR82fXuz5z0x5Ujl7Asfw1DlvXT9fiLmhgxYoetiTM"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/items": {
      "get": {
        "summary": "items",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "summary": "new-items",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "description": "",
            "required": true,
            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJjNmMyMjYwOC0wZDY1LTQ1YmUtYWNkZC00NTA4MzNiNjJiMGUiLCJyb2xlIjoiVXNlciIsImlhdCI6MTc1NTQ4ODkxMSwiZXhwIjoxNzU1NDg5ODExLCJqdGkiOiJBQW5qNjBtb1hiZE14eTlTTzdyN0dnIn0.5BT8sVgvg9Spv-0Nnv7br3058tAwCAuTsVCSld4ktTU",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "description": "",
            "required": true,
            "example": "TlyvbFv3QLmon6LOnBHjQ",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "description": "",
            "required": true,
            "example": "TlyvbFv3QLmon6LOnBHjQ",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "title",
                  "description"
                ]
              },
              "example": {
                "title": "First",
                "description": "demo"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "owner_id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "owner_id",
                    "title",
                    "description",
                    "created_at",
                    "updated_at"
                  ]
                },
                "example": {
                  "id": "d96275cc-f0b0-459b-a1b3-1d2620ee3101",
                  "owner_id": "c6c22608-0d65-45be-acdd-450833b62b0e",
                  "title": "First",
                  "description": "demo",
                  "created_at": "2025-08-18T03:57:27.661782Z",
                  "updated_at": "2025-08-18T03:57:27.661782Z"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {},
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "bearer": []
    }
  ]
}